refactor: moving mongo classes to internal.db.mongo
