+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/db/mongo/MongoCommandWithDistance.java
@@ -1,4 +1,4 @@
-package org.evomaster.client.java.controller.internal.db;
+package org.evomaster.client.java.controller.internal.db.mongo;
 
 public class MongoCommandWithDistance {
