+++ client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/db/mongo/MongoHandler.java
@@ -1,4 +1,4 @@
-package org.evomaster.client.java.controller.internal.db;
+package org.evomaster.client.java.controller.internal.db.mongo;
 
 import org.evomaster.client.java.controller.api.dto.database.execution.MongoFailedQuery;
 import org.evomaster.client.java.controller.api.dto.database.execution.MongoExecutionsDto;
